Refatora e corrige lógica de cálculo de custos no modal
Esta alteração resolve um problema de dependência circular e implementa a lógica de cálculo de custos correta no modal de nova encomenda, conforme as especificações do usuário.

- As funções de cálculo de custo (`calcular_materiais`, `calcular_custo_minimo`) foram movidas de `cogs/encomendas.py` para `cogs/calculator.py` para centralizar a lógica e resolver uma dependência circular.
- O modal agora usa `calcular_custo_minimo` para determinar o "Custo Mínimo de Fabricação", que é baseado no custo dos materiais crus (minérios).
- O "Valor de Venda Mínimo" é calculado separadamente, multiplicando a quantidade do item pelo seu preço de venda direto da API, obtido via `find_price`.
