+++ ui/modals.py
@@ -1,7 +1,7 @@
 import discord
 from ui.embeds import ConfirmView
 
-from cogs.calculator import calcular_custo_craft, find_price
+from cogs.calculator import find_price, calcular_custo_minimo
 
 class NewOrder(discord.ui.Modal):
     def __init__(self, bot, button_data, receitas, precos, produto_predefinido: str = None):
@@ -35,7 +35,7 @@ class NewOrder(discord.ui.Modal):
             quantidade = 1
 
         # CÁLCULO DE CUSTO DE CRAFT
-        preco_min, preco_max = calcular_custo_craft(
+        preco_min = calcular_custo_minimo(
             produto_name,
             quantidade,
             self.receitas,
@@ -52,12 +52,11 @@ class NewOrder(discord.ui.Modal):
 
         embed = discord.Embed(title='Confirmar nova encomenda!', color=discord.Colour.random())
 
-        if preco_min == 0 and preco_max == 0 and produto_name in self.receitas:
-            preco_min_str = preco_max_str = "❓"
+        if preco_min == 0 and produto_name in self.receitas:
+            preco_min_str = "❓"
             embed.set_footer(text="Custo zerado. Verifique se todos os materiais base possuem preço.")
         else:
             preco_min_str = f"$ {preco_min}"
-            preco_max_str = f"$ {preco_max}"
 
         embed.add_field(name='🧑 Nome', value=f'```{self.name.value}```', inline=False)
         embed.add_field(name='🕊️ Pombo', value=f'```{self.pombo.value}```', inline=False)
